DomusMercatoris/AI/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
numpy
pillow
scikit-learn
pillow-heif
